# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Documentation",
    "version": "1.0.0",
    "description": "This is the OpenAPI specification for your project."
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
    "responses": {
      "400": {
        "description": "Bad Request",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Invalid request parameters"
                },
                "code": {
                  "type": "string",
                  "example": "400"
                }
              }
            }
          }
        }
      },
...
      "500": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "An unexpected error occurred"
                },
                "code": {
                  "type": "string",
                  "example": "500"
                }
              }
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/crop-trait": {
      "get": {
        "operationId": "get-crop-trait",
        "summary": "Get crop traits\r",
        "description": "Returns all traits associated with a given crop",
        "tags": [
          "Crop-trait"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "crop",
            "schema": {
              "type": "string"
            },
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "number"
                      },
                      "traitName": {
                        "type": "string"
                      },
                      "entity": {
                        "type": "string"
                      },
                      "methodDescription": {
                        "type": "string"
                      },
                      "unit": {
                        "type": "string"
                      },
                      "min": {
                        "type": "number",
                        "nullable": true
                      },
                      "max": {
                        "type": "number",
                        "nullable": true
                      },
                      "variable": {
                        "type": "string"
                      },
                      "crop": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "id",
                      "traitName",
                      "entity",
                      "methodDescription",
                      "unit",
                      "variable",
                      "crop"
                    ]
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/crops": {
      "get": {
        "operationId": "get-crops",
        "summary": "Get Crops\r",
        "description": "Returns crops with optional filters, including related common names, species, product types, and physiological stages.",
        "tags": [
          "Crops"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "commonName",
            "schema": {
              "type": "string",
              "description": "Filter crops by common name"
            },
            "required": false,
            "description": "Filter crops by common name"
          },
          {
            "in": "query",
            "name": "species",
            "schema": {
              "type": "string",
              "description": "Filter crops by species name"
            },
            "required": false,
            "description": "Filter crops by species name"
          },
          {
            "in": "query",
            "name": "productType",
            "schema": {
              "type": "string",
              "description": "Filter crops by product type"
            },
            "required": false,
            "description": "Filter crops by product type"
          },
          {
            "in": "query",
            "name": "physiologicalStage",
            "schema": {
              "type": "string",
              "description": "Filter crops by physiological stage"
            },
            "required": false,
            "description": "Filter crops by physiological stage"
          }
        ],
        "responses": {
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/nir-models": {
      "get": {
        "operationId": "get-nir-models",
        "summary": "Get NIR Models\r",
        "description": "Returns NIR models with optional filters (manufacturer, type, name).",
        "tags": [
          "Nir-models"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "manufacturer",
            "schema": {
              "type": "string",
              "description": "Manufacturer of the NIR model"
            },
            "required": false,
            "description": "Manufacturer of the NIR model"
          },
          {
            "in": "query",
            "name": "type",
            "schema": {
              "type": "string",
              "description": "Type of the NIR model"
            },
            "required": false,
            "description": "Type of the NIR model"
          },
          {
            "in": "query",
            "name": "name",
            "schema": {
              "type": "string",
              "description": "Name of the NIR model"
            },
            "required": false,
            "description": "Name of the NIR model"
          }
        ],
        "responses": {
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/nirs-data": {
      "get": {
        "operationId": "get-nirs-data",
        "summary": "Get NIRS data per sample\r",
        "description": "Returns spectral NIR data grouped by sampleId",
        "tags": [
          "Nirs-data"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "crop",
            "schema": {
              "type": "string",
              "description": "Crop name"
            },
            "required": false,
            "description": "Crop name"
          },
          {
            "in": "query",
            "name": "qualityLab",
            "schema": {
              "type": "string",
              "description": "Quality lab name"
            },
            "required": false,
            "description": "Quality lab name"
          },
          {
            "in": "query",
            "name": "nirModel",
            "schema": {
              "type": "string",
              "description": "NIR model name"
            },
            "required": false,
            "description": "NIR model name"
          },
          {
            "in": "query",
            "name": "species",
            "schema": {
              "type": "string",
              "description": "Species name"
            },
            "required": false,
            "description": "Species name"
          },
          {
            "in": "query",
            "name": "plantingYear",
            "schema": {
              "type": "string",
              "description": "Planting year (YYYY)"
            },
            "required": false,
            "description": "Planting year (YYYY)"
          },
          {
            "in": "query",
            "name": "sampleDate",
            "schema": {
              "type": "string",
              "description": "Sample collection date (YYYY-MM-DD)"
            },
            "required": false,
            "description": "Sample collection date (YYYY-MM-DD)"
          },
          {
            "in": "query",
            "name": "physiologicalStage",
            "schema": {
              "type": "string",
              "description": "Physiological stage name"
            },
            "required": false,
            "description": "Physiological stage name"
          },
          {
            "in": "query",
            "name": "productType",
            "schema": {
              "type": "string",
              "description": "Product type name"
            },
            "required": false,
            "description": "Product type name"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/physiological-stages": {
      "get": {
        "operationId": "get-physiological-stages",
        "summary": "Get Physiological Stages\r",
        "description": "Returns physiological stages, optionally filtered by crop name.",
        "tags": [
          "Physiological-stages"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "crop",
            "schema": {
              "type": "string",
              "description": "Filter stages by crop name"
            },
            "required": false,
            "description": "Filter stages by crop name"
          }
        ],
        "responses": {
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
    "/species": {
      "get": {
        "operationId": "get-species",
        "summary": "Get Species\r",
        "description": "Returns species, optionally filtered by crop name.",
        "tags": [
          "Species"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "crop",
            "schema": {
              "type": "string",
              "description": "Filter species by crop name"
            },
            "required": false,
            "description": "Filter species by crop name"
          }
        ],
        "responses": {
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
  }
}